feat(mlcache) new [temporary] kong.mlcache module
This module is to replace the current database_cache.lua module with a
nicer and more powerful interface. This module is currently being
developed as a separate, generic lua-resty library and hence no tests
from this library were inserted into Kong's test suite.

This module is to disappear in favor of a dependency to the said
lua-resty-mlcache library (once completed).